完成

+++ src/main/java/nju/dao/ServiceDao.java
@@ -3,6 +3,6 @@ package nju.dao;
 import nju.model.ServiceData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceDao extends JpaRepository<ServiceData, Long> {
+public interface ServiceDao extends JpaRepository<ServiceData, Integer> {
 
 }
