+++ src/main/java/nju/dao/StaffDao.java
@@ -1,8 +1,11 @@
 package nju.dao;
 
 import nju.model.StaffData;
+import nju.xyf.request.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffDao extends JpaRepository<StaffData, Long> {
+import java.util.List;
+
+public interface StaffDao extends JpaRepository<StaffData, Integer> {
 
 }
