+++ src/main/java/nju/dao/StoreDao.java
@@ -3,6 +3,6 @@ package nju.dao;
 import nju.model.StoreData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreDao extends JpaRepository<StoreData, Long> {
+public interface StoreDao extends JpaRepository<StoreData, Integer> {
 
 }
